+++ tests/dispatch.php
@@ -1,0 +1,6 @@
+<?php
+
+function dispatch(object $object)
+{
+    Butler\Guru\Tests\TestCase::$dispatched[] = $object;
+}